Removed "Test" tab from custom uploader settings form and instead moved tabs to response form
