+++ ShareX.UploadersLib/Forms/CustomUploaderSettingsForm.cs
@@ -36,7 +36,6 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
@@ -98,9 +97,6 @@ namespace ShareX.UploadersLib
             rtbResultThumbnailURL.AddContextMenu();
             rtbResultDeletionURL.AddContextMenu();
             rtbResultErrorMessage.AddContextMenu();
-            rtbResult.AddContextMenu();
-            rtbResponseInfo.AddContextMenu();
-            rtbResponseText.AddContextMenu();
             eiCustomUploaders.ObjectType = typeof(CustomUploaderItem);
             CustomUploaderAddDestinationTypes();
             cbRequestMethod.Items.AddRange(Enum.GetNames(typeof(HttpMethod)));
@@ -668,9 +664,6 @@ namespace ShareX.UploadersLib
 
             btnImageUploaderTest.Enabled = btnTextUploaderTest.Enabled = btnFileUploaderTest.Enabled =
                 btnURLShortenerTest.Enabled = btnURLSharingServiceTest.Enabled = false;
-            rtbResult.ResetText();
-            rtbResponseText.ResetText();
-            rtbResponseInfo.ResetText();
             lbCustomUploaderList.SelectedIndex = index;
 
             CustomUploaderItem item = Config.CustomUploadersList[index];
@@ -726,91 +719,12 @@ namespace ShareX.UploadersLib
             {
                 if (result != null)
                 {
-                    StringBuilder sbResult = new StringBuilder();
-
-                    if (((type == CustomUploaderDestinationType.ImageUploader || type == CustomUploaderDestinationType.TextUploader ||
-                        type == CustomUploaderDestinationType.FileUploader) && !string.IsNullOrEmpty(result.URL)) ||
-                        (type == CustomUploaderDestinationType.URLShortener && !string.IsNullOrEmpty(result.ShortenedURL)) ||
-                        (type == CustomUploaderDestinationType.URLSharingService && !result.IsError && !string.IsNullOrEmpty(result.URL)))
-                    {
-                        if (!string.IsNullOrEmpty(result.ShortenedURL))
-                        {
-                            sbResult.AppendLine("Shortened URL: " + result.ShortenedURL);
-                        }
-
-                        if (!string.IsNullOrEmpty(result.URL))
-                        {
-                            sbResult.AppendLine("URL: " + result.URL);
-                        }
-
-                        if (!string.IsNullOrEmpty(result.ThumbnailURL))
-                        {
-                            sbResult.AppendLine("Thumbnail URL: " + result.ThumbnailURL);
-                        }
-
-                        if (!string.IsNullOrEmpty(result.DeletionURL))
-                        {
-                            sbResult.AppendLine("Deletion URL: " + result.DeletionURL);
-                        }
-                    }
-                    else if (result.IsError)
-                    {
-                        sbResult.AppendLine(result.ErrorsToString());
-                    }
-                    else
-                    {
-                        sbResult.AppendLine(Resources.UploadersConfigForm_TestCustomUploader_Error__Result_is_empty_);
-                    }
-
-                    rtbResult.Text = sbResult.ToString();
-
-                    if (result.ResponseInfo != null)
-                    {
-                        rtbResponseText.ResetText();
-                        rtbResponseText.Text = result.ResponseInfo.ResponseText;
-
-                        UpdateResponseInfoTextBox(result.ResponseInfo, true);
-                    }
-
-                    tcCustomUploader.SelectedTab = tpTest;
+                    ResponseForm responseForm = new ResponseForm(result);
+                    responseForm.Show();
                 }
 
                 btnImageUploaderTest.Enabled = btnTextUploaderTest.Enabled = btnFileUploaderTest.Enabled =
                     btnURLShortenerTest.Enabled = btnURLSharingServiceTest.Enabled = true;
-            }
-        }
-
-        private void UpdateResponseInfoTextBox(ResponseInfo responseInfo, bool includeResponseText)
-        {
-            rtbResponseInfo.ResetText();
-
-            rtbResponseInfo.SetFontBold();
-            rtbResponseInfo.AppendText("Status code:\r\n");
-            rtbResponseInfo.SetFontRegular();
-            rtbResponseInfo.AppendText($"({(int)responseInfo.StatusCode}) {responseInfo.StatusDescription}");
-
-            if (!string.IsNullOrEmpty(responseInfo.ResponseURL))
-            {
-                rtbResponseInfo.SetFontBold();
-                rtbResponseInfo.AppendText("\r\n\r\nResponse URL:\r\n");
-                rtbResponseInfo.SetFontRegular();
-                rtbResponseInfo.AppendText(responseInfo.ResponseURL);
-            }
-
-            if (responseInfo.Headers != null && responseInfo.Headers.Count > 0)
-            {
-                rtbResponseInfo.SetFontBold();
-                rtbResponseInfo.AppendText("\r\n\r\nHeaders:\r\n");
-                rtbResponseInfo.SetFontRegular();
-                rtbResponseInfo.AppendText(responseInfo.Headers.ToString().TrimEnd('\r', '\n'));
-            }
-
-            if (includeResponseText && !string.IsNullOrEmpty(responseInfo.ResponseText))
-            {
-                rtbResponseInfo.SetFontBold();
-                rtbResponseInfo.AppendText("\r\n\r\nResponse text:\r\n");
-                rtbResponseInfo.SetFontRegular();
-                rtbResponseInfo.AppendText(responseInfo.ResponseText);
             }
         }
 
@@ -1310,54 +1224,6 @@ namespace ShareX.UploadersLib
             CustomUploaderItem uploader = CustomUploaderGetSelected();
             if (uploader != null) uploader.ErrorMessage = rtbResultErrorMessage.Text;
             CustomUploaderSyntaxHighlight(rtbResultErrorMessage);
-        }
-
-        private void txtCustomUploaderLog_LinkClicked(object sender, LinkClickedEventArgs e)
-        {
-            URLHelpers.OpenURL(e.LinkText);
-        }
-
-        private void tsbCustomUploaderJSONFormat_Click(object sender, EventArgs e)
-        {
-            string response = rtbResponseText.Text;
-            if (!string.IsNullOrEmpty(response))
-            {
-                try
-                {
-                    response = Helpers.JSONFormat(response, Formatting.Indented);
-                    rtbResponseText.Text = response;
-                }
-                catch
-                {
-                    MessageBox.Show("Formatting failed.", "ShareX", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
-            }
-        }
-
-        private void tsbCustomUploaderXMLFormat_Click(object sender, EventArgs e)
-        {
-            string response = rtbResponseText.Text;
-            if (!string.IsNullOrEmpty(response))
-            {
-                try
-                {
-                    response = Helpers.XMLFormat(response);
-                    rtbResponseText.Text = response;
-                }
-                catch
-                {
-                    MessageBox.Show("Formatting failed.", "ShareX", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
-            }
-        }
-
-        private void tsbCustomUploaderCopyResponseText_Click(object sender, EventArgs e)
-        {
-            string response = rtbResponseText.Text;
-            if (!string.IsNullOrEmpty(response))
-            {
-                ClipboardHelpers.CopyText(response);
-            }
         }
 
         private void cbCustomUploaderImageUploader_SelectedIndexChanged(object sender, EventArgs e)
